baseball_talent_manager/backend/requirements.txt
alembic==1.14.0
argon2-cffi==25.1.0
fastapi==0.115.5
passlib==1.7.4
pydantic[email]==2.12.5
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
python_jose==3.3.0
SQLAlchemy==2.0.36
uvicorn==0.32.1